+++ lib/fb_news_attachments_video.dart
@@ -39,8 +39,10 @@ class _FbNewsAttachmentsVideoState extends State<FbNewsAttachmentsVideo> {
       allowFullScreen: false,
       showControls: true,
       customControls: MaterialControls(),
-      autoInitialize: true,
+      autoInitialize: false,
     );
+
+    await _chewieController.videoPlayerController.initialize();
     setState(() {});
   }
 
@@ -51,7 +53,7 @@ class _FbNewsAttachmentsVideoState extends State<FbNewsAttachmentsVideo> {
           ? _chewieController.videoPlayerController.value.size.height
           : null,
       child: _chewieController != null &&
-              _chewieController.videoPlayerController.value.initialized
+              _chewieController.videoPlayerController.value.isInitialized
           ? Chewie(
               controller: _chewieController,
             )
@@ -59,8 +61,6 @@ class _FbNewsAttachmentsVideoState extends State<FbNewsAttachmentsVideo> {
               mainAxisAlignment: MainAxisAlignment.center,
               children: const [
                 CircularProgressIndicator(),
-                SizedBox(height: 20),
-                Text('Loading...'),
               ],
             ),
     );